requirements.txt
streamlit
yfinance
google-generativeai
pandas
requests
beautifulsoup4